tool to evaluate an agent on a scramble
I just used this tool to evaluate a network I trained overnight. It had 128 hidden units and was trained with episode length 8 and batch size 1024. It got up to an average reward of 3.8, meaning that it solved roughly 4.5 pieces on average.

After running the above network on a few scrambles, it was clear that the network tries to solve pieces all around the cube rather than solving isolated chunks. It is most likely easier to solve spread-out pieces, especially when you only have 8 moves anyway.